+++ ballista/scheduler/Cargo.toml
@@ -59,6 +59,7 @@ futures = { workspace = true }
 graphviz-rust = { version = "0.9", optional = true }
 http = "1.1"
 log = { workspace = true }
+md-5 = { version = "^0.10.0" }
 object_store = { workspace = true }
 once_cell = { version = "1.16.0", optional = true }
 parking_lot = { workspace = true }
